feat: Move log query access LogQueryAccessRole
Moves the role to a separate construct, keeping the stack responsible
only for adding constructs/resources.
